+++ packages/custom_lint_core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Upgraded `analyzer` to `>=5.7.0 <5.11.0`
 - `LintRuleNodeRegistry` and other AstVisitor-like now are based off `GeneralizingAstVisitor` instead of `GeneralizingAstVisitor`
+- Exposes the Pubspec in CustomLintContext
 
 ## 0.3.2 - 2023-03-09